Changed everything to use roles instead of capabilities, Rewrote the whole logic and used a code sniffer
